Bug 1428740 - Improve Marionette unit tests for parent and content crashes. r=maja_zf
For parent process crashes a Wait().until() condition was added, which is
not necessary because if the parent (chrome) process dies, the browser
window will be gone immediately, and the IOError is thrown directly.

Further the tests should check that the correct type of crash is shown
in the IOError exceptions.

MozReview-Commit-ID: JC90CObZQ3r

UltraBlame original commit: 59bb34f20076d62dcce516e05878e781b17960f5

+++ testing/marionette/harness/marionette_harness/tests/unit/test_crash.py
@@ -3,6 +3,7 @@
 
 
 import glob
+import os
 import shutil
 
 from marionette_driver import Wait
@@ -42,7 +43,8 @@ class MockMozCrash(object):
             except TimeoutException:
                 minidump_files = []
 
-            shutil.rmtree(dump_directory)
+            if os.path.isdir(dump_directory):
+                shutil.rmtree(dump_directory)
 
             return len(minidump_files)
         else:
@@ -105,13 +107,8 @@ class BaseCrashTestCase(MarionetteTestCase):
 class TestCrash(BaseCrashTestCase):
 
     def test_crash_chrome_process(self):
-        with self.assertRaises(IOError):
-            self.crash(chrome=True)
-            Wait(self.marionette, timeout=self.socket_timeout,
-                 ignored_exceptions=NoSuchWindowException).until(
-                lambda _: self.marionette.get_url(),
-                message="Expected IOError exception for content crash not raised."
-            )
+        self.assertRaisesRegexp(IOError, "Process crashed",
+                                self.crash, chrome=True)
 
         
         self.assertNotIn(self.marionette.instance.runner.returncode, (None, 0))
@@ -132,8 +129,7 @@ class TestCrash(BaseCrashTestCase):
         
         
         
-        
-        with self.assertRaises(IOError):
+        with self.assertRaisesRegexp(IOError, "Content process crashed"):
             self.crash(chrome=False)
             Wait(self.marionette, timeout=self.socket_timeout,
                  ignored_exceptions=NoSuchWindowException).until(
@@ -164,13 +160,8 @@ class TestCrashInSetUp(BaseCrashTestCase):
     def setUp(self):
         super(TestCrashInSetUp, self).setUp()
 
-        with self.assertRaises(IOError):
-            self.crash(chrome=True)
-            Wait(self.marionette, timeout=self.socket_timeout,
-                 ignored_exceptions=NoSuchWindowException).until(
-                lambda _: self.marionette.get_url(),
-                message="Expected IOError exception for content crash not raised."
-            )
+        self.assertRaisesRegexp(IOError, "Process crashed",
+                                self.crash, chrome=True)
 
         
         self.assertNotIn(self.marionette.instance.runner.returncode, (None, 0))
@@ -187,13 +178,8 @@ class TestCrashInTearDown(BaseCrashTestCase):
 
     def tearDown(self):
         try:
-            with self.assertRaises(IOError):
-                self.crash(chrome=True)
-                Wait(self.marionette, timeout=self.socket_timeout,
-                     ignored_exceptions=NoSuchWindowException).until(
-                    lambda _: self.marionette.get_url(),
-                    message="Expected IOError exception for content crash not raised."
-                )
+            self.assertRaisesRegexp(IOError, "Process crashed",
+                                    self.crash, chrome=True)
 
             
             self.assertNotIn(self.marionette.instance.runner.returncode, (None, 0))
